Refactored util functions
