+++ package.json
@@ -29,7 +29,10 @@
     "CHANGELOG.md",
     "LICENSE",
     "openapi-format.js",
+    "util-casing.js",
+    "util-filter.js",
     "util-log-output.js",
+    "util-sort.js",
     "util-types.js",
     "defaultSort.json",
     "defaultFilter.json",
